Finished making many tedious changes to the halo finder, so there are no global variables.  But it compiles!  Removed unnecessary I/O files for the inline finder.
--HG--
branch : trunk

+++ src/enzo/FOF_indexx.C
@@ -1,10 +1,14 @@
+#include "macros_and_parameters.h"
+#include "typedefs.h"
+#include "global_data.h"
+
 #define NRANSI
-#include "nrutil.h"
+#include "FOF_nrutil.h"
 #define SWAP(a,b) itemp=(a);(a)=(b);(b)=itemp;
 #define M 7
 #define NSTACK 50
 
-void indexx(unsigned long n, float arr[], int indx[])
+void indexx(int n, float arr[], int indx[])
 {
 	int i,indxt,ir=n,itemp,j,k,l=1;
 	int jstack=0,*istack;
